Extract items custom fields controller actions to a concern
